[React/RN] Move shortenAdress util to react-core
