+++ packages/react-native/src/evm/index.ts
@@ -42,7 +42,6 @@ export { Web3Button } from "./components/Web3Button";
 
 // utilities
 export * from "./utils/uri";
-export * from "./utils/addresses";
 export { createSyncStorage as createLocalStorage } from "../core/AsyncStorage";
 export { createAsyncLocalStorage } from "../core/AsyncStorage";
 export { createSecureStorage } from "../core/SecureStorage";
